commit second

+++ Services/EmpRegisterImpl.cs
@@ -1,14 +1,39 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.Win32;
+using Project_ASP.Net_And_Angular.Helper;
 using Project_ASP.Net_And_Angular.Models;
 using System.Diagnostics;
+using System.Globalization;
 
 namespace Project_ASP.Net_And_Angular.Services;
 
 public class EmpRegisterImpl : IEmpRegister
 {
     private DatabaseContext db;
-    public EmpRegisterImpl(DatabaseContext db) { this.db = db; }
+    private IConfiguration configuration;
+    public EmpRegisterImpl(DatabaseContext db, IConfiguration configuration)
+    {
+        this.db = db;
+        this.configuration = configuration;
+    }
+
+   
+    public bool checkExistEmailAndUsername(string email, string username)
+    {
+        try
+        {
+            var debug = db.EmpRegisters.Where(e => e.Username == username && e.Email == email).Count();
+            Debug.WriteLine("------------" + debug);
+            return db.EmpRegisters.Where(e => e.Username == username && e.Email == email).Count() > 0;
+
+        }
+        catch (Exception ex)
+        {
+            Debug.WriteLine(ex);
+            return false;
+        }
+    }
+
     public bool create(EmpRegister register)
     {
         try
@@ -27,8 +52,16 @@ public class EmpRegisterImpl : IEmpRegister
     {
         try
         {
-            db.EmpRegisters.Remove(db.EmpRegisters.Find(id));
-            return db.SaveChanges() > 0;
+            if(db.PoliciesonEmployees.Where(pe=>pe.Empno== id).Count() > 0 || 
+                db.PolicyRequestDetails.Where(pe => pe.Empno == id).Count() > 0)
+            {
+                return false;
+            }
+            else
+            {
+                db.EmpRegisters.Remove(db.EmpRegisters.Find(id));
+                return db.SaveChanges() > 0;
+            }
         }
         catch (Exception ex)
         {
@@ -41,7 +74,7 @@ public class EmpRegisterImpl : IEmpRegister
     {
         try
         {
-            return db.EmpRegisters.Where(e => e.Email == email).Count()>0;
+            return db.EmpRegisters.Where(e => e.Email == email).Count() > 0;
         }
         catch (Exception ex)
         {
@@ -67,11 +100,11 @@ public class EmpRegisterImpl : IEmpRegister
     {
         try
         {
-            return db.EmpRegisters.Where(e=>e.Email == emailNoTracking).Select(e => new
+            return db.EmpRegisters.Where(e => e.Email == emailNoTracking).Select(e => new
             {
                 empNo = e.Empno,
                 designation = e.Designation,
-                joihDate = e.Joindate,
+                joinDate = e.Joindate.Value.ToString("dd-MM-yyyy"),
                 salary = e.Salary,
                 firstName = e.Firstname,
                 lastName = e.Lastname,
@@ -82,8 +115,6 @@ public class EmpRegisterImpl : IEmpRegister
                 state = e.State,
                 country = e.Country,
                 city = e.City,
-                policyStatus = e.Policystatus,
-                policyId = e.Policyid,
                 roleName = e.Rolename,
                 accountStatus = e.AccountStatus,
                 securityCode = e.SecurityCode,
@@ -105,7 +136,7 @@ public class EmpRegisterImpl : IEmpRegister
             {
                 empNo = e.Empno,
                 designation = e.Designation,
-                joihDate = e.Joindate,
+                joinDate = e.Joindate.Value.ToString("dd-MM-yyyy"),
                 salary = e.Salary,
                 firstName = e.Firstname,
                 lastName = e.Lastname,
@@ -116,8 +147,6 @@ public class EmpRegisterImpl : IEmpRegister
                 state = e.State,
                 country = e.Country,
                 city = e.City,
-                policyStatus = e.Policystatus,
-                policyId = e.Policyid,
                 roleName = e.Rolename,
                 accountStatus = e.AccountStatus,
                 securityCode = e.SecurityCode,
@@ -139,19 +168,17 @@ public class EmpRegisterImpl : IEmpRegister
             {
                 empNo = e.Empno,
                 designation = e.Designation,
-                joihDate = e.Joindate,
+                joinDate = e.Joindate.Value.ToString("dd-MM-yyyy"),
                 salary = e.Salary,
                 firstName = e.Firstname,
                 lastName = e.Lastname,
                 username = e.Username,
-                password = e.Password,
+ //               password = e.Password,
                 address = e.Address,
                 contactNo = e.Contactno,
                 state = e.State,
                 country = e.Country,
                 city = e.City,
-                policyStatus = e.Policystatus,
-                policyId = e.Policyid,
                 roleName = e.Rolename,
                 accountStatus = e.AccountStatus,
                 securityCode = e.SecurityCode,
@@ -165,6 +192,32 @@ public class EmpRegisterImpl : IEmpRegister
         }
     }
 
+    public bool forgotPas(string email, string username)
+    {
+        try
+        {
+            var check = db.EmpRegisters.Where(e => e.Username == username && e.Email == email).Count() >0;
+            if(check) {
+                var data = db.EmpRegisters.Where(e => e.Username == username && e.Email == email).AsNoTracking().SingleOrDefault();
+                string content = RandomHelper.RandomInt(6);
+                var mailHelper = new MailHelper(configuration);
+                var contentInput = "Your activation code is <h3>" + content + "</h3>";
+                var statusSendEmail = mailHelper.Send(configuration["Gmail:UserName"], data.Email, "Security Alert", contentInput);
+                data.SecurityCode = content;
+                db.Entry(data).State = Microsoft.EntityFrameworkCore.EntityState.Modified;
+                return db.SaveChanges() > 0;
+            }
+            else
+            {
+                return false;
+            }
+           
+        }catch(Exception ex) { 
+        Debug.WriteLine(ex);
+            return false;
+        }
+    }
+
     public dynamic get()
     {
         try
@@ -173,7 +226,7 @@ public class EmpRegisterImpl : IEmpRegister
             {
                 empNo = e.Empno,
                 designation = e.Designation,
-                joihDate = e.Joindate,
+                joinDate =e.Joindate.Value.ToString("dd-MM-yyyy"),
                 salary = e.Salary,
                 firstName = e.Firstname,
                 lastName = e.Lastname,
@@ -184,8 +237,6 @@ public class EmpRegisterImpl : IEmpRegister
                 state = e.State,
                 country = e.Country,
                 city = e.City,
-                policyStatus = e.Policystatus,
-                policyId = e.Policyid,
                 roleName = e.Rolename,
                 accountStatus = e.AccountStatus,
                 securityCode = e.SecurityCode,
@@ -203,23 +254,44 @@ public class EmpRegisterImpl : IEmpRegister
     {
         try
         {
-            var data = db.EmpRegisters.Where(e => e.Username == userName).FirstOrDefault();
-            if(BCrypt.Net.BCrypt.Verify(data.Password,password))
+            var check = db.EmpRegisters.Where(e => e.Username == userName).SingleOrDefault();
+            Debug.WriteLine(BCrypt.Net.BCrypt.Verify(password, check.Password));
+            if (BCrypt.Net.BCrypt.Verify(password, check.Password))
             {
                 return new
                 {
-                    username = data.Username,
-                    role = data.Rolename
+                    id = check.Empno,
+                    username = check.Username,
+                    role = check.Designation,
+                    status = check.AccountStatus
                 };
             }
             else
             {
-                return new
-                {
-                    result = "Password invalid"
-                };
+                return null;
             }
-        }catch(Exception ex)
+        }
+        catch (Exception ex)
+        {
+            Debug.WriteLine(ex);
+            return null;
+        }
+    }
+
+    public bool sendSecurityCode(string email)
+    {
+        try
+        {
+            var data = db.EmpRegisters.Where(e => e.Email == email).AsNoTracking().SingleOrDefault();
+            string content = RandomHelper.RandomInt(6);
+            var mailHelper = new MailHelper(configuration);
+            var contentInput = "Your activation code is <h3>" + content + "</h3>";
+            var statusSendEmail = mailHelper.Send(configuration["Gmail:UserName"], data.Email, "Security Alert", contentInput);
+            data.SecurityCode = content;
+            db.Entry(data).State = Microsoft.EntityFrameworkCore.EntityState.Modified;
+            return db.SaveChanges() > 0;
+        }
+        catch (Exception ex)
         {
             Debug.WriteLine(ex);
             return false;
@@ -232,6 +304,67 @@ public class EmpRegisterImpl : IEmpRegister
         {
             db.Entry(register).State = Microsoft.EntityFrameworkCore.EntityState.Modified;
             return db.SaveChanges() > 0;
+        }
+        catch (Exception ex)
+        {
+            Debug.WriteLine(ex);
+            return false;
+        }
+    }
+
+    public bool verify(string code, string email)
+    {
+        try
+        {
+            var data = db.EmpRegisters.Where(e => e.Email == email).AsNoTracking().SingleOrDefault();
+            if (data.SecurityCode == code)
+            {
+                data.SecurityCode = null;
+                data.AccountStatus = true;
+                db.Entry(data).State = Microsoft.EntityFrameworkCore.EntityState.Modified;
+                return db.SaveChanges() > 0;
+            }
+            else
+            {
+                return false;
+            }
+
+        }
+        catch (Exception ex)
+        {
+            Debug.WriteLine(ex);
+            return false;
+        }
+    }
+
+    public bool changeForgotPass(int id, string newPass)
+    {
+        try
+        {
+            var data = db.EmpRegisters.Find(id);
+            data.Password = newPass;
+            db.Entry(data).State = Microsoft.EntityFrameworkCore.EntityState.Modified;
+            return db.SaveChanges() > 0;
+        }
+        catch(Exception ex)
+        {
+            Debug.WriteLine(ex);
+            return false;
+        }
+    }
+
+    public bool updateNoSavePass(int id, EmpRegister register)
+    {
+       
+        try
+        {
+            var data = db.EmpRegisters.Find(id);
+            string pass = data.Password.ToString();
+            data = register;
+            data.Password = pass;
+            data.Empno= id;
+            db.Entry(data).State= EntityState.Modified;
+            return db.SaveChanges()>0;
         }
         catch (Exception ex)
         {
