+++ Services/PolicyImpl.cs
@@ -12,10 +12,12 @@ public class PolicyImpl : IPolicy
         try
         {
             db.Policies.Add(policy);
-            return db.SaveChanges()>0;
-        }catch(Exception ex) {
+            return db.SaveChanges() > 0;
+        }
+        catch (Exception ex)
+        {
             Debug.WriteLine(ex);
-            return false; 
+            return false;
         }
     }
 
@@ -24,8 +26,20 @@ public class PolicyImpl : IPolicy
 
         try
         {
-            db.Policies.Remove(db.Policies.Find(id));
-            return db.SaveChanges() > 0;
+            if (db.PoliciesonEmployees.Where(p => p.Policyid == id).Count() > 0 ||
+               db.PolicyRequestDetails.Where(p => p.PolicyId == id).Count() > 0 ||
+               db.PolicyApprovalDetails.Where(p => p.PolicyId == id).Count() > 0 ||
+               db.PolicyTotalDescriptions.Where(p => p.Policyid == id).Count() > 0
+                )
+            {
+                return false;
+            }
+            else
+            {
+                db.Policies.Remove(db.Policies.Find(id));
+                return db.SaveChanges() > 0;
+            }
+
         }
         catch (Exception ex)
         {
@@ -38,7 +52,7 @@ public class PolicyImpl : IPolicy
     {
         try
         {
-            return db.Policies.Where(p=>p.Policyid == id).Select(p => new
+            return db.Policies.Where(p => p.Policyid == id).Select(p => new
             {
                 policyId = p.Policyid,
                 policyName = p.Policyname,
@@ -61,10 +75,10 @@ public class PolicyImpl : IPolicy
     {
         try
         {
-            return db.Policies.Select(p=> new
+            return db.Policies.Select(p => new
             {
                 policyId = p.Policyid,
-                policyName= p.Policyname,
+                policyName = p.Policyname,
                 policyDesc = p.Policydesc,
                 amount = p.Amount,
                 emi = p.Emi,
@@ -72,7 +86,8 @@ public class PolicyImpl : IPolicy
                 companyName = p.Company.CompanyName,
                 medicalid = p.Medicalid
             }).ToList();
-        }catch(Exception e)
+        }
+        catch (Exception e)
         {
             Debug.WriteLine(e);
             return false;
